[ new ] mixed induction & coinduction (#3)
Also a big refactoring of the test suite to run tests based on feature sets rather than
where in the library they were defined. This allows us to verify that the behaviour on
shared feature sets is uniform across all variants of the library i.e. that the addition of
new features does not introduce regressions.